+++ config/expert/CodinDice/code/CGCode.cpp
@@ -1,0 +1,23 @@
+#include <iostream>
+#include <string>
+#include <vector>
+#include <algorithm>
+
+using namespace std;
+
+int main()
+{
+    int n;
+    cin >> n; cin.ignore();
+    for (int i = 0; i < n; i++) {
+        int x;
+        int y;
+        string face;
+        cin >> x >> y >> face; cin.ignore();
+    }
+
+    // Write an answer using cout. DON'T FORGET THE "<< endl"
+    // To debug: cerr << "Debug messages..." << endl;
+
+    cout << "answer" << endl;
+}