+++ config/expert/CodinDice/code/CGCode.ts
@@ -1,0 +1,12 @@
+const N: number = parseInt(readline());
+for (let i = 0; i < N; i++) {
+    var inputs: string[] = readline().split(' ');
+    const x: number = parseInt(inputs[0]);
+    const y: number = parseInt(inputs[1]);
+    const face: string = inputs[2];
+}
+
+// Write an answer using console.log()
+// To debug: console.error('Debug messages...');
+
+console.log('answer');