+++ CHANGELOG.md
@@ -33,6 +33,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - configuration for "Hourglass".
 - configuration for "The Crime Scene".
 - configuration for "Dungeon Designer".
+- configuration for "CodinDice".
 
 ## [1.16.0] - 2025-06-30
 ### Added